+++ Actions/RunPipeline/RunPipeline.ps1
@@ -193,6 +193,7 @@ try {
     }
 
     # Replace secret names in install.apps and install.testApps
+    $tempDependenciesLocation = Join-Path $ENV:RUNNER_TEMP "Dependencies-$(Get-Random)"
     foreach($list in @('Apps','TestApps')) {
         $install."$list" = @($install."$list" | ForEach-Object {
             $pattern = '.*(\$\{\{\s*([^}]+?)\s*\}\}).*'
@@ -203,22 +204,44 @@ try {
             else {
                 $finalUrl = $url
             }
-            # Check validity of URL
+
             if ($finalUrl -like 'http*://*') {
+                # Check validity of URL
                 try {
                     Invoke-WebRequest -Method Head -UseBasicParsing -Uri $finalUrl | Out-Null
-                }
-                catch {
+                } catch {
                     throw "Setting: install$($list) contains an inaccessible URL: $($url). Error was: $($_.Exception.Message)"
                 }
+
+                # Try downloading the app file
+                try {
+                    if (-not (Test-Path $tempDependenciesLocation)) {
+                        New-Item -Path $tempDependenciesLocation -ItemType Directory | Out-Null
+                    }
+                    $urlWithoutQuery = $finalUrl.Split('?')[0].TrimEnd('/')
+                    $rawFileName = [System.IO.Path]::GetFileName($urlWithoutQuery)
+                    $decodedFileName = [Uri]::UnescapeDataString($rawFileName)
+                    $appFile = Join-Path $tempDependenciesLocation $decodedFileName
+                    Invoke-WebRequest -Method GET -UseBasicParsing -Uri $finalUrl -OutFile $appFile -MaximumRetryCount 3 -RetryIntervalSec 5 | Out-Null
+                }
+                catch {
+                    throw "Could not download app from URL: $($url). Error was: $($_.Exception.Message)"
+                }
+            } else {
+                $appFile = $_
             }
-            return $finalUrl
+            return $appFile
         })
     }
 
     # Analyze app.json version dependencies before launching pipeline
 
     # Analyze InstallApps and InstallTestApps before launching pipeline
+    if ((-not $settings.doNotPublishApps)) {
+        # Test that InstallApps are not symbols packages
+        Import-Module (Join-Path -Path $PSScriptRoot -ChildPath ".\RunPipeline.psm1" -Resolve)
+        Test-InstallApps -AllInstallApps ($install.Apps + $install.TestApps) -ProjectPath $projectPath
+    }
 
     # Check if codeSignCertificateUrl+Password is used (and defined)
     if (!$settings.doNotSignApps -and $codeSignCertificateUrl -and $codeSignCertificatePassword -and !$settings.keyVaultCodesignCertificateName) {