+++ RELEASENOTES.md
@@ -29,6 +29,7 @@ Please note that some automated features are premium and require the use of [Git
 - Issue 1937 trackALAlertsInGitHub is failing in preview
 - DeployTo settings from environment-specific AL-Go settings are not applied when deploying
 - `ReadSettings` action outputs too much information that is mainly used for debugging
+- Issue 1512: Throw a warning before trying to publish symbols packages
 
 ## v7.3